Some coding style changes, making the data readers and writers more uniform. No functional change.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@1715 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/multiple_data_writers_controller.h
@@ -33,15 +33,15 @@ public:
 
   bool ready();
 
-  Data_writer_ptr get_data_writer(size_t i);
+  Data_writer_ptr get_data_writer(unsigned int i);
 
-	// This is the set of listening IP/port
-	void get_listening_ip(std::vector<uint64_t>& ip_port);
+  // This is the set of listening IP/port
+  void get_listening_ip(std::vector<uint64_t>& ip_port);
 
 private:
   void add_data_writer(unsigned int i, Data_writer_ptr writer);
 
-  std::vector< Data_writer_ptr >  data_writers;
+  std::vector<Data_writer_ptr> data_writers;
 
   TCP_Connection tcp_connection;
 };
